refactor(app): extract text-manipulation orchestrators to text_ops (phase 3d)
Move smart_home, smart_home_visual_line, toggle_comment, and
goto_matching_bracket out of render.rs into a new app/text_ops.rs sibling.

These are cursor/edit operations whose pure decision logic for smart-home
already lives in app/smart_home.rs (extracted in phase 1h); the
orchestrators that gather buffer state, build events, and apply them to
the active buffer move here.

render.rs goes from 3514 -> 3051 lines (-463). 2026 tests still pass.
